requirements.txt
asgiref==3.8.1
Django==4.0.10
djangorestframework==3.13.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
